+++ Compiler/src/edu/mit/compilers/st/ArrayDesc.java
@@ -1,9 +1,9 @@
 package edu.mit.compilers.st;
 
 public class ArrayDesc extends Descriptor {
-    private Integer cap;
+    private Long cap;
 
-    public ArrayDesc(String type, String text, Integer cap) {
+    public ArrayDesc(String type, String text, Long cap) {
         super(type, text);
         this.cap = cap;
     }
@@ -11,7 +11,7 @@ public class ArrayDesc extends Descriptor {
     // find if index is overbound
     @Override
     public final String findVar(String idxStr) {
-        int idx = Integer.parseInt(idxStr);
+        Long idx = Long.parseLong(idxStr);
         if (idx < cap) {
             return "ok";
         }
@@ -23,7 +23,7 @@ public class ArrayDesc extends Descriptor {
         return null;
     }
 
-    public final Integer getCap() {
+    public final Long getCap() {
         return this.cap;
     }
 }
